Stremio Trakt Addon + Mise à jour de la stack Stremio

+++ includes/functions.sh
@@ -791,17 +791,39 @@ function suppression_appli() {
     manage_account_yml sub.immich " "
     ;;
   streamfusion)
-    sudo rm -rf ${SETTINGS_STORAGE}/docker/${USER}/${APPSELECTED}
-    docker rm -f zilean streamfusion >/dev/null 2>&1
-    # Il faut gérer les DB postgres dans la fonction 'check_and_remove_shared_containers'
-    check_and_remove_shared_containers ${APPSELECTED}
-    docker volume prune -f >/dev/null 2>&1
+    docker rm -f warp streamfusion >/dev/null 2>&1
+    if [ $DELETE -eq 1 ]; then
+        sudo rm -rf ${SETTINGS_STORAGE}/docker/${USER}/${APPSELECTED}
+        docker volume rm warp-data >/dev/null 2>&1
+        # Il faut gérer les DB postgres dans la fonction 'check_and_remove_shared_containers'
+        check_and_remove_shared_containers ${APPSELECTED}
+        docker volume prune -f >/dev/null 2>&1
+    fi
     ;;
   stremiocatalogs)
-    sudo rm -rf ${SETTINGS_STORAGE}/docker/${USER}/${APPSELECTED}
     docker rm -f ${APPSELECTED} >/dev/null 2>&1
-    # Il faut gérer les DB postgres dans la fonction 'check_and_remove_shared_containers'
-    check_and_remove_shared_containers ${APPSELECTED}
+    if [ $DELETE -eq 1 ]; then
+        sudo rm -rf ${SETTINGS_STORAGE}/docker/${USER}/${APPSELECTED}
+        # Il faut gérer les DB postgres dans la fonction 'check_and_remove_shared_containers'
+        check_and_remove_shared_containers ${APPSELECTED}
+    fi
+    ;;
+  stremiotrakt)
+    docker rm -f ${APPSELECTED} >/dev/null 2>&1
+    if [ $DELETE -eq 1 ]; then
+        sudo rm -rf ${SETTINGS_STORAGE}/docker/${USER}/${APPSELECTED}
+        # Il faut gérer les DB postgres dans la fonction 'check_and_remove_shared_containers'
+        check_and_remove_shared_containers ${APPSELECTED}
+    fi
+    ;;
+  zilean)
+    docker rm -f zilean >/dev/null 2>&1
+    if [ $DELETE -eq 1 ]; then
+        sudo rm -rf ${SETTINGS_STORAGE}/docker/${USER}/${APPSELECTED}
+        # Il faut gérer les DB postgres dans la fonction 'check_and_remove_shared_containers'
+        check_and_remove_shared_containers ${APPSELECTED}
+        docker volume prune -f >/dev/null 2>&1
+    fi
     ;;
   esac
 
@@ -838,12 +860,13 @@ EOF
 
 function check_and_remove_shared_containers() {
   local app="$1"
-  local other_apps=("zilean" "streamfusion" "stremiocatalogs")
-  local other_apps_running=false
+  # Stremio Base Removal
+  local stremio_apps=("zilean" "streamfusion" "stremiocatalogs" "stremiotrakt")
+  local stremio_apps_running=false
 
-  for other_app in "${other_apps[@]}"; do
-    if [ "$other_app" != "$app" ] && docker ps -q --filter name="$other_app" | grep -q .; then
-      other_apps_running=true
+  for stremio_app in "${stremio_apps[@]}"; do
+    if [ "$stremio_app" != "$app" ] && docker ps -q --filter name="$stremio_app" | grep -q .; then
+      stremio_apps_running=true
       break
     fi
   done
@@ -851,11 +874,16 @@ function check_and_remove_shared_containers() {
   if docker ps -a --filter "name=stremio-postgres" --format "{{.Names}}" | grep -q "stremio-postgres"; then
       case "$app" in
         streamfusion)
-          docker exec -e PGPASSWORD=stremio stremio-postgres psql -U stremio -d postgres -c "DROP DATABASE IF EXISTS \"zilean\";" || echo "Failed to drop zilean database."
           docker exec -e PGPASSWORD=stremio stremio-postgres psql -U stremio -d postgres -c "DROP DATABASE IF EXISTS \"streamfusion\";" || echo "Failed to drop streamfusion database."
           ;;
         stremiocatalogs)
           docker exec -e PGPASSWORD=stremio stremio-postgres psql -U stremio -d postgres -c "DROP DATABASE IF EXISTS \"stremio-catalog-db\";" || echo "Failed to drop stremio-catalog-db database."
+          ;;
+        stremiotrakt)
+          docker exec -e PGPASSWORD=stremio stremio-postgres psql -U stremio -d postgres -c "DROP DATABASE IF EXISTS \"stremio-trakt-db\";" || echo "Failed to drop stremio-trakt-db database."
+          ;;
+        zilean)
+          docker exec -e PGPASSWORD=stremio stremio-postgres psql -U stremio -d postgres -c "DROP DATABASE IF EXISTS \"zilean\";" || echo "Failed to drop zilean database."
           ;;
         *)
           docker exec -e PGPASSWORD=stremio stremio-postgres psql -U stremio -d postgres -c "DROP DATABASE IF EXISTS \"$app-db\";" || echo "Failed to drop $app-db database."
@@ -863,9 +891,10 @@ function check_and_remove_shared_containers() {
       esac
   fi
 
-  if ! $other_apps_running; then
+  if ! $stremio_apps_running; then
     docker rm -f stremio-postgres stremio-redis >/dev/null 2>&1 || echo "Failed to remove containers."
-    docker volume rm data-stremio-postgres data-stremio-redis >/dev/null 2>&1 || echo "Failed to remove volumes."
+    sudo rm -rf ${SETTINGS_STORAGE}/docker/${USER}/stremio-redis >/dev/null 2>&1
+    sudo rm -rf ${SETTINGS_STORAGE}/docker/${USER}/stremio-postgres >/dev/null 2>&1
   fi
 }
